Moved SettingsContainer to .utils
