+++ django_replicated/utils.py
@@ -21,3 +21,19 @@ class Routers(object):
 
 
 routers = Routers()
+
+
+class SettingsContainer(object):
+    def __init__(self):
+        from django.conf import settings as django_settings
+        from . import settings as default_settings
+
+        default_settings_names = dir(default_settings)
+        django_settings_names = dir(django_settings)
+
+        for k in default_settings_names:
+            if k in django_settings_names:
+                new_value = getattr(django_settings, k)
+            else:
+                new_value = getattr(default_settings, k)
+            setattr(self, k, new_value)
